+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/accounts/components/form/main.js
@@ -543,8 +543,22 @@ define(['config'], function(Config) {
             if (saved !== this.saved) {
                 var type = (!!this.options.data && !!this.options.data.id) ? 'edit' : 'add';
                 this.sandbox.emit('sulu.header.toolbar.state.change', type, saved, true);
+                this.propagateState(saved);
             }
             this.saved = saved;
+        },
+
+        /**
+         * Propagates the state of the content with an event
+         *  sulu.content.saved when the content has been saved
+         *  sulu.content.changed when the content has been changed
+         */
+        propagateState: function(saved) {
+            if (!!saved) {
+                this.sandbox.emit('sulu.content.saved');
+            } else {
+                this.sandbox.emit('sulu.content.changed');
+            }
         },
 
         listenForChange: function() {