+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/contacts/components/form/main.js
@@ -185,8 +185,8 @@ define(['config'], function(Config) {
                 // add collection filters to form
                 this.sandbox.emit('sulu.contact-form.add-collectionfilters', form);
                 this.sandbox.form.setData(form, data).then(function() {
-
-                    if(!!startForm){
+                    // prevents emitting events when form is not there anymore
+                    if (!!startForm) {
                         this.sandbox.start(form);
                     } else {
                         this.sandbox.start('#contact-fields');
@@ -280,6 +280,7 @@ define(['config'], function(Config) {
                     this.options.data = data;
                     this.initContactData();
                     this.setFormData(data);
+                    this.setTitle();
                     this.setHeaderBar(true);
                 }, this);
 
@@ -414,8 +415,22 @@ define(['config'], function(Config) {
                 if (saved !== this.saved) {
                     var type = (!!this.options.data && !!this.options.data.id) ? 'edit' : 'add';
                     this.sandbox.emit('sulu.header.toolbar.state.change', type, saved, true);
+                    this.propagateState(saved);
                 }
                 this.saved = saved;
+            },
+
+            /**
+             * Propagates the state of the content with an event
+             *  sulu.content.saved when the content has been saved
+             *  sulu.content.changed when the content has been changed
+             */
+            propagateState: function(saved) {
+                if (!!saved) {
+                    this.sandbox.emit('sulu.content.saved');
+                } else {
+                    this.sandbox.emit('sulu.content.changed');
+                }
             },
 
             /**