+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/contacts/main.js
@@ -287,6 +287,7 @@ define([
                         // TODO update address lists
                         this.sandbox.emit('sulu.contacts.contacts.saved', model);
                     } else {
+                        this.sandbox.emit('sulu.content.saved');
                         this.sandbox.emit('sulu.router.navigate', 'contacts/contacts/edit:' + model.id + '/details');
                     }
                 }.bind(this),